Move functional envtest to separate package
As functional tests should never depend on unexported names from the
system under test this patch moves the functional tests to a separate
package. Also we took the opportunity to gather test helpers to a
separate file.
